Feature/tables created

+++ src/main/java/learning/platform/entity/User.java
@@ -14,14 +14,14 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String email;
     private String password;
     private Role role;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
